Fix CORS issue, use gunicorn for flask
