+++ frontend/vue.config.js
@@ -1,11 +1,27 @@
 module.exports = {
   lintOnSave: false,
   devServer: {
+    stats: {
+      colors: true,
+      hash: false,
+      version: false,
+      timings: false,
+      assets: false,
+      chunks: false,
+      modules: false,
+      reasons: false,
+      children: false,
+      source: false,
+      errors: false,
+      errorDetails: false,
+      warnings: false,
+      publicPath: false
+    },
     proxy: {
       '/socket.io': {
         target: `http://${process.env.VUE_APP_FLASK_HOST}:5000`,
         ws: true,
-        changeOrigin: true,
+        changeOrigin: true
       },
     },
   },
